made some changes to the files.
I updated the mystery icon image to the correct name, so now the mystery box shows up on the opponent side. I also changed the sizing of the rock, paper, scissors images and mystery box to better fit the page. i also changed the color of the page though trial and error of different color combonations.